Enable WASM for macOS builds (#57)
* Enable WASM for macOS builds

This changes the release workflow to enable wasm on MacOS envoy
builds. To do so, this updates the minimum MacOS version to 14.5,
which corresponds to the upstream minimum, and allows access to
newer APIs like `aligned_alloc`. Until we do this, we have no
nice way to use MacOS without docker, and with such a small change
needed, we should just do it.

Signed-off-by: Adrian Cole <adrian@tetrate.io>

* dry

Signed-off-by: Adrian Cole <adrian@tetrate.io>

---------

Signed-off-by: Adrian Cole <adrian@tetrate.io>
Co-authored-by: Adrian Cole <adrian@tetrate.io>